# password-manager-api-swagger.yaml
openapi: 3.0.0
info:
  title: Password Management API
  description: API for securely managing passwords, including storing, viewing, editing, and sharing passwords between users and groups.
  version: 1.0.0
  contact:
    name: Project Team
    email: contact@passwordmanager.com
tags:
  - name: user
    description: User related operations
  - name: group
    description: Group related operations
  - name: password
    description: Password related operations
  - name: log
    description: Log related operations

paths:
  /users:
    post:
      summary: Create a new user
      description: Adds a new user to the system.
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input (e.g. missing fields, invalid data types)
        '409':
          description: Conflict (e.g. user already exists)
        '500':
          description: Internal server error

    get:
      summary: Get all users
      description: Retrieves a list of all users.
      tags:
        - user
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Internal server error

  /users/{userId}:
    get:
      summary: Get user by ID
      description: Retrieves a user's details by their ID.
      tags:
        - user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
        '500':
          description: Internal server error

    put:
      summary: Update user
      description: Updates the information of a specific user.
      tags:
        - user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated successfully
        '400':
          description: Invalid input (e.g. invalid data type)
        '404':
          description: User not found
        '500':
          description: Internal server error

    delete:
      summary: Delete user
      description: Deletes a user from the system by their ID.
      tags:
        - user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted successfully
        '404':
          description: User not found
        '500':
          description: Internal server error

  /groups:
    post:
      summary: Create a new group
      description: Creates a new group to share passwords with members.
      tags:
        - group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid input (e.g. missing fields)
        '500':
          description: Internal server error

    get:
      summary: Get all groups
      description: Retrieves a list of all groups.
      tags:
        - group
      responses:
        '200':
          description: List of all groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '500':
          description: Internal server error

  /groups/{groupId}:
    get:
      summary: Get group by ID
      description: Retrieves a group's details by its ID.
      tags:
        - group
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Group found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          description: Group not found
        '500':
          description: Internal server error

    put:
      summary: Update group
      description: Updates the information of a specific group.
      tags:
        - group
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
      responses:
        '200':
          description: Group updated successfully
        '400':
          description: Invalid input (e.g. invalid data type)
        '404':
          description: Group not found
        '500':
          description: Internal server error

    delete:
      summary: Delete group
      description: Deletes a group from the system by its ID.
      tags:
        - group
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Group deleted successfully
        '404':
          description: Group not found
        '500':
          description: Internal server error

  /passwords:
    post:
      summary: Create a new password entry
      description: Adds a new password entry for a user.
      tags:
        - password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Password'
      responses:
        '201':
          description: Password created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Password'
        '400':
          description: Invalid input (e.g. missing fields, incorrect password format)
        '500':
          description: Internal server error

    get:
      summary: Get all passwords
      description: Retrieves all passwords for the authenticated user.
      tags:
        - password
      responses:
        '200':
          description: List of all passwords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Password'
        '500':
          description: Internal server error

  /passwords/{passwordId}:
    get:
      summary: Get password entry by ID
      description: Retrieves a password entry by its ID.
      tags:
        - password
      parameters:
        - in: path
          name: passwordId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Password found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Password'
        '404':
          description: Password not found
        '500':
          description: Internal server error

    put:
      summary: Update password entry
      description: Updates a specific password entry.
      tags:
        - password
      parameters:
        - in: path
          name: passwordId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Password'
      responses:
        '200':
          description: Password updated successfully
        '400':
          description: Invalid input (e.g. invalid password format)
        '404':
          description: Password not found
        '500':
          description: Internal server error

    delete:
      summary: Delete password entry
      description: Deletes a password entry by its ID.
      tags:
        - password
      parameters:
        - in: path
          name: passwordId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Password deleted successfully
        '404':
          description: Password not found
        '500':
          description: Internal server error

  /logs:
    get:
      summary: Get all access logs
      description: Retrieves a list of all access logs.
      tags:
        - log
      responses:
        '200':
          description: List of all logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Log'
        '500':
          description: Internal server error

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        password:
          type: string
        groupId:
          type: integer

    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        members:
          type: array
          items:
            type: integer
        passwords:
          type: array
          items:
            type: integer

    Password:
      type: object
      properties:
        id:
          type: integer
        urlOrDescription:
          type: string
        username:
          type: string
        password:
          type: string
        metadata:
          type: object
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time

    Log:
      type: object
      properties:
        userId:
          type: integer
        actionType:
          type: string
        actionDate:
          type: string
          format: date-time